+++ spec/badgeville/reward_spec.rb
@@ -1,6 +1,10 @@
 require 'spec_helper'
 
 describe Badgeville::Reward do
+  before(:each) do
+    Badgeville.configure(private_key: 'PRIVATE_KEY')
+  end
+
   let(:reward_json) do
     {
       'name'             => 'A Way with Words',
@@ -41,6 +45,77 @@ describe Badgeville::Reward do
       'active'           => true,
       'message'          => 'You won the internet!'
     }
+  end
+
+  describe 'find_by_player' do
+    before(:each) do
+      stub_request(:get, /.*\/rewards\.json/).to_return(
+        status: 200,
+        body: {data: [reward_json]}.to_json
+      )
+    end
+
+    it 'finds every reward for the player by the player_id' do
+      Badgeville::Reward.find_by_player({player_id: 'PLAYER_ID'})
+
+      a_request(:get, /.*\/rewards\.json/).with(
+        query: {'player_id' => 'PLAYER_ID', 'page' => '1', 'per_page' => '50', 'include_totals' => 'true'}
+      ).should have_been_made
+    end
+
+    it 'finds every reward for the player by the email and site' do
+      Badgeville::Reward.find_by_player({site: 'example.org', email: 'user@example.org'})
+
+      a_request(:get, /.*\/rewards\.json/).with(
+        query: {'site' => 'example.org', 'email' => 'user@example.org', 'page' => '1', 'per_page' => '50', 'include_totals' => 'true'}
+      ).should have_been_made
+    end
+
+    it 'returns an empty array if no rewards could be found' do
+      stub_request(:get, /.*\/rewards\.json/).to_return(
+        status: 200,
+        body: {data: []}.to_json
+      )
+
+      rewards = Badgeville::Reward.find_by_player({player_id: 'PLAYER_ID'})
+
+      rewards.should be_empty
+    end
+
+    it 'returns an array of reward objects' do
+      rewards = Badgeville::Reward.find_by_player({player_id: 'PLAYER_ID'})
+      rewards.first.should be_a(Badgeville::Reward)
+      rewards.first.name.should == 'A Way with Words'
+    end
+
+    it 'gets ALL the rewards with pagination' do
+      stub_request(:get, /.*\/rewards\.json/).to_return(
+        {body: {data: [reward_json], paging: {current_page: 1, total_pages: 3}}.to_json},
+        {body: {data: [reward_json.merge(name: 'Typinghero')], paging: {current_page: 2, total_pages: 3}}.to_json},
+        {body: {data: [reward_json.merge(name: 'Eduard Khil')], paging: {current_page: 3, total_pages: 3}}.to_json},
+      )
+
+      rewards = Badgeville::Reward.find_by_player({player_id: 'PLAYER_ID'})
+
+      rewards.should have(3).items
+      rewards[0].name.should == 'A Way with Words'
+      rewards[1].name.should == 'Typinghero'
+      rewards[2].name.should == 'Eduard Khil'
+    end
+
+    it 'raises an ArgumentError when player_id or site and email are not provided' do
+      -> {
+        Badgeville::Reward.find_by_player({email: 'user@example.org'})
+      }.should raise_error(ArgumentError, 'You have to provide a player_id or a site and email')
+
+      -> {
+        Badgeville::Reward.find_by_player({site: 'example.org'})
+      }.should raise_error(ArgumentError, 'You have to provide a player_id or a site and email')
+    end
+
+    it 'handles some errors' do
+      pending('TODO: Handle all errors. First find out how Badgeville responds')
+    end
   end
 
   context "when json is a reward definition" do